server/requirements.txt
fastapi
uvicorn
psutil
python-dotenv
requests
pyyaml
